Доработаны сервисы. Логика перенесена в сервисы из контроллеров. Вместо null выбрасываются исключения и добавлен обработчик исключений. С учетом этого доработаны тесты сервисов. Также осуществлено логирование. Тесты контроллеров пока не доделаны.

+++ src/main/java/com/example/ProjectManager/service/impl/ProjectServiceImpl.java
@@ -6,6 +6,7 @@ import com.example.ProjectManager.model.User;
 import com.example.ProjectManager.repository.ProjectRepository;
 import com.example.ProjectManager.repository.TaskRepository;
 import com.example.ProjectManager.repository.UserRepository;
+import com.example.ProjectManager.Exceptions.NotFoundException;
 import com.example.ProjectManager.service.ProjectService;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
@@ -27,6 +28,7 @@ public class ProjectServiceImpl implements ProjectService {
     private final UserRepository userRepository;
 
     private static final String NAME_ADMIN = "admin";
+    private static final String PROJECT_NOT_FOUND = "Такого проекта не существует";
 
     @Override
     public List<Project> findAllProjects() {
@@ -46,17 +48,36 @@ public class ProjectServiceImpl implements ProjectService {
                 .getAuthentication().getPrincipal();
         Optional<User> currentUser = userRepository.findByName(userDetails.getUsername());
         currentUser.ifPresent(project::setUser);
-        return projectRepository.save(project);
+        Project correctProject;
+        //теперь если передан project без id, то randomId создается тут (раньше было в классе сущности)
+        if (project.getId() == null) {
+            correctProject = Project.builder()
+                    .id(UUID.randomUUID())
+                    .name(project.getName())
+                    .description(project.getDescription())
+                    .endDate(project.getEndDate())
+                    .user(project.getUser())
+                    .tasks(project.getTasks())
+                    .build();
+        } else {
+            correctProject = project;
+        }
+        return projectRepository.save(correctProject);
     }
 
     @Override
-    public Optional<Project> findById(UUID id) {
+    public Optional<Project> findById(UUID id) throws NotFoundException {
+        if (projectRepository.findById(id).isEmpty()) {
+            throw new NotFoundException(PROJECT_NOT_FOUND);
+        }
         return projectRepository.findById(id);
     }
 
     @Override //изучить как работает hibernate
-    public Project updateProject(Project project) {
-        if (!projectRepository.findById(project.getId()).isPresent()) return null;
+    public Project updateProject(Project project) throws NotFoundException {
+        if (projectRepository.findById(project.getId()).isEmpty()) {
+            throw new NotFoundException(PROJECT_NOT_FOUND);
+        }
         Project newProject = projectRepository.findById(project.getId()).get(); //hibernate отслеживает new project
         newProject.setName(project.getName());
         newProject.setDescription(project.getDescription());
@@ -66,7 +87,10 @@ public class ProjectServiceImpl implements ProjectService {
 
     @Override
     @Transactional //изучить как работает
-    public void deleteProject(UUID id) {
+    public void deleteProject(UUID id) throws NotFoundException {
+        if (findById(id).isEmpty()) {
+            throw new NotFoundException(PROJECT_NOT_FOUND);
+        }
         taskRepository.deleteByProjectId(id);
         projectRepository.deleteById(id);
     }
