+++ src/main/java/com/example/ProjectManager/service/impl/TaskServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.ProjectManager.service.impl;
 
+import com.example.ProjectManager.Exceptions.NotFoundException;
 import com.example.ProjectManager.model.Task;
 import com.example.ProjectManager.repository.ProjectRepository;
 import com.example.ProjectManager.repository.TaskRepository;
@@ -18,6 +19,8 @@ import java.util.UUID;
 @AllArgsConstructor
 public class TaskServiceImpl implements TaskService {
 
+    private static final String PROJECT_NOT_EXIST = "Такого проекта не существует";
+    private static final String TASK_NOT_EXIST = "Такой задачи не существует";
     private final TaskRepository taskRepository;
     private final ProjectRepository projectRepository;
 
@@ -32,26 +35,49 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public List<Task> findAllTasksForProject(UUID projectId) {
+    public List<Task> findAllTasksForProject(UUID projectId) throws NotFoundException {
+        if (projectNotExist(projectId)) {
+            throw new NotFoundException(PROJECT_NOT_EXIST);
+        }
         return taskRepository.findByProjectId(projectId);
     }
 
     @Override
     @Transactional
-    public Task saveTask(Task task) {
-        Task currentTask = taskRepository.save(task);
-        projectRepository.findById(task.getProjectId()).get().getTasks().add(currentTask);
-        return currentTask;
+    public Task saveTask(Task task) throws NotFoundException {
+        UUID projectId = task.getProjectId();
+        if (projectRepository.findById(projectId).isEmpty()) {
+            throw new NotFoundException(PROJECT_NOT_EXIST);
+        }
+        Task correctTask;
+        if (task.getId() == null) {
+            correctTask = Task.builder()
+                    .projectId(task.getProjectId())
+                    .id(UUID.randomUUID())
+                    .description(task.getDescription())
+                    .endDate(task.getEndDate())
+                    .build();
+        } else {
+            correctTask = task;
+        }
+
+        projectRepository.findById(correctTask.getProjectId()).get().getTasks().add(correctTask);
+        return taskRepository.save(correctTask);
     }
 
     @Override
-    public Optional<Task> findById(UUID id) {
+    public Optional<Task> findById(UUID id) throws NotFoundException {
+        if (taskRepository.findById(id).isEmpty()) {
+            throw new NotFoundException(TASK_NOT_EXIST);
+        }
         return taskRepository.findById(id);
     }
 
     @Override
-    public Task updateTask(Task task) {
-        if (!taskRepository.findById(task.getId()).isPresent()) return null;
+    public Task updateTask(Task task) throws NotFoundException {
+        if (taskRepository.findById(task.getId()).isEmpty()) {
+            throw new NotFoundException(TASK_NOT_EXIST);
+        }
         Task newTask = taskRepository.findById(task.getId()).get();
         newTask.setName(task.getName());
         newTask.setDescription(task.getDescription());
@@ -60,7 +86,10 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public void deleteTask(UUID id) {
+    public void deleteTask(UUID id) throws NotFoundException {
+        if (taskRepository.findById(id).isEmpty()) {
+            throw new NotFoundException(TASK_NOT_EXIST);
+        }
         taskRepository.deleteById(id);
     }
 }
